+++ public/images/index2.jsx
@@ -1,55 +1,56 @@
 const images = {
-  home: require('./home.png').default,
-  logo: require('./logo.png'),
-  collection: require('./paint-palette.png'),
-  remove: require('./magic-wand1.png'),
-  category: require('./folder.png'),
-  picture: require('./picture.png'),
-  discount: require('./discount-coupon.png'),
-  remove: require('./magic-wand1.png'),
-  category: require('./folder.png'),
-  defaultAvatar: require('./default-avatar.png'),
-  balance: require('./banknotes.png'),
-  eCoin: require('./coin.png'),
-  design: require('./database.png'),
-  order: require('./online-order.png'),
-  collection2: require('./paint-roller (1).png'),
-  transaction: require('./history.png'),
-  renew: require('./repeater.png'),
-  painter: require('./painter.png'),
-  };
-  
+  home: require("./home.png").default,
+  logo: require("./logo.png"),
+  collection: require("./paint-palette.png"),
+  remove: require("./magic-wand1.png"),
+  category: require("./folder.png"),
+  picture: require("./picture.png"),
+  discount: require("./discount-coupon.png"),
+  remove: require("./magic-wand1.png"),
+  category: require("./folder.png"),
+  defaultAvatar: require("./default-avatar.png"),
+  balance: require("./banknotes.png"),
+  eCoin: require("./coin.png"),
+  design: require("./database.png"),
+  order: require("./online-order.png"),
+  collection2: require("./paint-roller (1).png"),
+  transaction: require("./history.png"),
+  renew: require("./repeater.png"),
+  painter: require("./painter.png"),
+};
+
 export const proImages = [
-  require('./pro1.jpg'),
-  require('./pro2.jpg'),
-  require('./pro3.jpg'),
-  require('./img4.jpg'),
-  ];
-    
+  require("./pro1.jpg"),
+  require("./pro2.jpg"),
+  require("./pro3.jpg"),
+  require("./img4.jpg"),
+];
+
 export const designIcon = {
-  paperRolls: require('./paper-rolls.png').default,
-  presentation: require('./presentation.svg'),
-  picture: require('./picture.png'),
-  poster: require('./poster.svg'),
-  instagram: require('./instagram.svg'),
-  facebook: require('./facebook.svg'),
-  docs: require('./docs.svg'),
-  youtube: require('./youtube.svg'),
+  paperRolls: require("./paper-rolls.png").default,
+  presentation: require("./presentation.svg"),
+  picture: require("./picture.png"),
+  poster: require("./poster.svg"),
+  instagram: require("./instagram.svg"),
+  facebook: require("./facebook.svg"),
+  docs: require("./docs.svg"),
+  youtube: require("./youtube.svg"),
 };
 
 export const designAction = {
-  edit: require('./edit.png').default,
-  copy: require('./copy.png'),
-  delete: require('./delete.png'),
-  eye: require('./eye.png'),
-  share: require('./share.png'),
-}
+  edit: require("./edit.png").default,
+  copy: require("./copy.png"),
+  delete: require("./delete.png"),
+  eye: require("./eye.png"),
+  share: require("./share.png"),
+  download: require("./download.png"),
+};
 
 export const contactInfo = {
-  man: require('./contact-man.png'),
-  gmail: require('./gmail.png'),
-  phone: require('./phone-call.png'),
-  bg: require('./bg-designer.png')
-}
+  man: require("./contact-man.png"),
+  gmail: require("./gmail.png"),
+  phone: require("./phone-call.png"),
+  bg: require("./bg-designer.png"),
+};
 
 export default images;
