+++ src/app/(main)/(home)/endow/page.jsx
@@ -1,8 +1,8 @@
-import EndowProduct from '@/components/Slide/EndowProduct'
-import React from 'react'
+"use client";
+
+import EndowProduct from "@/components/Slide/EndowProduct";
+import React from "react";
 
 export default function Page() {
-  return (
-    <EndowProduct />
-  )
+  return <EndowProduct />;
 }
