+++ src/app/(main)/(home)/event/page.jsx
@@ -1,8 +1,8 @@
-import EventProductSlider from '@/components/Slide/EventProductSlider'
-import React from 'react'
+"use client";
+
+import EventProductSlider from "@/components/Slide/EventProductSlider";
+import React from "react";
 
 export default function Page() {
-  return (
-    <EventProductSlider />
-  )
+  return <EventProductSlider />;
 }
