+++ src/app/(main)/project/cooking/page.jsx
@@ -1,10 +1,12 @@
-import EventProductSlider from '@/components/Slide/EventProductSlider'
-import React from 'react'
+"use client";
+
+import EventProductSlider from "@/components/Slide/EventProductSlider";
+import React from "react";
 
 export default function page() {
   return (
     <>
       <EventProductSlider />
     </>
-  )
+  );
 }
