+++ src/app/(main)/project/layout.jsx
@@ -1,13 +1,14 @@
+"use client";
 
-import RecommendAction from '@/components/RecommendPage/RecommendAction'
-import RecommendBanner from '@/components/RecommendPage/RecommendBanner'
-import NewProductsSlider from '@/components/Slide/NewproductsSlider'
-import React from 'react'
+import RecommendAction from "@/components/RecommendPage/RecommendAction";
+import RecommendBanner from "@/components/RecommendPage/RecommendBanner";
+import NewProductsSlider from "@/components/Slide/NewproductsSlider";
+import React from "react";
 
 export default function layout(props) {
   return (
-    <div className='flex-col w-[90%]'>
-      <div className='w-full pt-5'>
+    <div className="flex-col w-[90%]">
+      <div className="w-full pt-5">
         <RecommendBanner />
       </div>
       <div>
@@ -16,5 +17,5 @@ export default function layout(props) {
       {props.children}
       <NewProductsSlider />
     </div>
-  )
+  );
 }
