+++ src/app/(main)/project/more/page.jsx
@@ -1,8 +1,10 @@
-import CollectionProductSlider from '@/components/Slide/CollectionProductSlider'
-import EndowProduct from '@/components/Slide/EndowProduct'
-import EventProductSlider from '@/components/Slide/EventProductSlider'
-import SerisProductSlider from '@/components/Slide/SerisProductSlider'
-import React from 'react'
+"use client";
+
+import CollectionProductSlider from "@/components/Slide/CollectionProductSlider";
+import EndowProduct from "@/components/Slide/EndowProduct";
+import EventProductSlider from "@/components/Slide/EventProductSlider";
+import SerisProductSlider from "@/components/Slide/SerisProductSlider";
+import React from "react";
 
 export default function page() {
   return (
@@ -12,5 +14,5 @@ export default function page() {
       <EndowProduct />
       <EventProductSlider />
     </>
-  )
+  );
 }
