+++ src/app/(main)/project/recommend/page.jsx
@@ -1,10 +1,12 @@
-import EndowProduct from '@/components/Slide/EndowProduct'
-import React from 'react'
+"use client";
+
+import EndowProduct from "@/components/Slide/EndowProduct";
+import React from "react";
 
 export default function page() {
   return (
     <>
       <EndowProduct />
     </>
-  )
+  );
 }
