+++ src/components/Slide/EndowProduct.jsx
@@ -1,12 +1,11 @@
-import React from "react";
 import DefaultSlide from "./DefaultSlide";
 import { getLogoProductApi } from "@/api/product";
 
 const EndowProduct = () => {
-
   return (
     <DefaultSlide
       apiAction={async () => {
+        // "use server";
         const products = await getLogoProductApi({
           category_id: "12",
           limit: "12",
