+++ src/components/Slide/EventProductSlider.jsx
@@ -1,12 +1,12 @@
-import React from "react";
+
 import DefaultSlide from "./DefaultSlide";
 import { getLogoProductApi } from "@/api/product";
 
 const EventProductSlider = () => {
-
   return (
     <DefaultSlide
       apiAction={async () => {
+        // "use server";
         const products = await getLogoProductApi({
           category_id: "9",
           limit: "12",
