+++ src/components/Slide/GiftProductSlider.jsx
@@ -1,3 +1,4 @@
+import React from "react";
 import { getLogoProductApi } from "@/api/product";
 import DefaultSlide from "./DefaultSlide";
