+++ src/components/Slide/NewproductsSlider.jsx
@@ -1,12 +1,11 @@
-import React from "react";
 import DefaultSlide from "./DefaultSlide";
 import { getNewProducts } from "@/api/product";
 
 const NewProductsSlider = () => {
-
   return (
     <DefaultSlide
       apiAction={async () => {
+        // "use server";
         const products = await getNewProducts();
         return products;
       }}
