+++ src/components/Slide/SerisProductSlider.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import DefaultSlide from "./DefaultSlide";
 import { getserisProductApi } from "@/api/product";
 
@@ -6,6 +5,7 @@ const SerisProductSlider = () => {
   return (
     <DefaultSlide
       apiAction={async () => {
+        // "use server";
         const response = await getserisProductApi({ limit: "12", page: "1" });
         const data = { listData: [...response.data] };
         return data; // Return the fetched products
