+++ src/components/YourProduct/DefaultPage.jsx
@@ -1,14 +1,15 @@
-import React, { useEffect, useState } from 'react';
-import ProductCard from '@/components/YourProduct/ProductCard';
-import { deleteProductAPI, duplicateProductAPI } from '@/api/product';
-import { checkTokenCookie } from '@/utils/cookie';
-import { toast } from 'react-toastify';
-import { Skeleton } from 'antd';
+import React, { useEffect, useState } from "react";
+import ProductCard from "@/components/YourProduct/ProductCard";
+import { deleteProductAPI, duplicateProductAPI } from "@/api/product";
+import { checkTokenCookie } from "@/utils/cookie";
+import { toast } from "react-toastify";
+import { Skeleton } from "antd";
 
 export default function DefaultPage({ getData }) {
   const [products, setProducts] = useState([]);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
+  const [uploadedImageUrl, setUploadedImageUrl] = useState(null);
 
   useEffect(() => {
     const fetchData = async () => {
@@ -29,11 +30,13 @@ export default function DefaultPage({ getData }) {
     try {
       await deleteProductAPI({
         token: checkTokenCookie(),
-        id: productId
+        id: productId,
       });
-      const updatedProducts = products.filter(product => product.id !== productId);
+      const updatedProducts = products.filter(
+        (product) => product.id !== productId
+      );
       setProducts(updatedProducts);
-      toast.success("Xóa thành công !!!")
+      toast.success("Xóa thành công !!!");
     } catch (error) {
       console.error("Error deleting product:", error.message);
     }
@@ -42,24 +45,48 @@ export default function DefaultPage({ getData }) {
   const onDuplicateProduct = async (productId) => {
     try {
       const newProduct = await duplicateProductAPI({
-         token: checkTokenCookie(),
-        id: productId
+        token: checkTokenCookie(),
+        id: productId,
       });
-      const updatedProducts = [newProduct, ...products]; 
+      const updatedProducts = [newProduct, ...products];
       setProducts(updatedProducts);
-      toast.success("Nhân bản thành công !!!")
+      toast.success("Nhân bản thành công !!!");
     } catch (error) {
       console.error("Error duplicating product:", error.message);
     }
   };
 
+  const onDownloadProduct = async (imageUrl) => {
+    try {
+      const response = await fetch(imageUrl);
+      const blob = await response.blob();
+
+      const url = URL.createObjectURL(blob);
+
+      const a = document.createElement("a");
+      a.href = url;
+      a.download = "downloaded_image.png"; // Set the desired filename
+
+      document.body.appendChild(a);
+      a.click();
+      document.body.removeChild(a);
+
+      URL.revokeObjectURL(url);
+      toast.success("Tải xuống thành công !!!");
+    } catch (error) {
+      console.error("Error downloading the image:", error);
+    }
+  };
+
   if (loading) {
-    return (<Skeleton
-              avatar
-              paragraph={{
-                rows: 10,
-              }}
-            />);
+    return (
+      <Skeleton
+        avatar
+        paragraph={{
+          rows: 10,
+        }}
+      />
+    );
   }
 
   if (error) {
@@ -67,10 +94,11 @@ export default function DefaultPage({ getData }) {
   }
 
   return (
-    <ProductCard 
-      products={products} 
-      onDeleteProduct={onDeleteProduct} 
-      onDuplicateProduct={onDuplicateProduct} 
-    />   
+    <ProductCard
+      products={products}
+      onDeleteProduct={onDeleteProduct}
+      onDownloadProduct={onDownloadProduct}
+      onDuplicateProduct={onDuplicateProduct}
+    />
   );
 }
