+++ src/components/YourProduct/ProductCard.jsx
@@ -1,17 +1,23 @@
-import Image from 'next/image';
-import React from 'react';
-import { designAction } from '../../../public/images/index2';
-import { Skeleton } from 'antd';
+import Image from "next/image";
+import React from "react";
+import { designAction } from "../../../public/images/index2";
+import { Skeleton } from "antd";
 
-export default function ProductCard({ products, onDeleteProduct, onDuplicateProduct }) {
+export default function ProductCard({
+  products,
+  onDeleteProduct,
+  onDownloadProduct,
+  onDuplicateProduct,
+}) {
   let buttonsData = [
     {
-      text: 'Sửa',
+      text: "Sửa",
       icon: designAction.edit,
-      action: (productId) => console.log('Edit action for product ID:', productId),
+      action: (productId) =>
+        console.log("Edit action for product ID:", productId),
     },
     {
-      text: 'Xóa',
+      text: "Xóa",
       icon: designAction.delete,
       action: onDeleteProduct,
     },
@@ -19,19 +25,21 @@ export default function ProductCard({ products, onDeleteProduct, onDuplicateProd
 
   if (onDuplicateProduct) {
     buttonsData.push({
-      text: 'Nhân bản',
+      text: "Nhân bản",
       icon: designAction.copy,
       action: onDuplicateProduct,
     });
   }
 
   return (
-    <div className={`w-[100%] mx-auto grid grid-cols-4 grid-flow-row ${!onDuplicateProduct ? 'gap-8' : 'gap-4'}`}>
+    <div
+      className={`w-[100%] mx-auto grid grid-cols-4 grid-flow-row ${
+        !onDuplicateProduct ? "gap-8" : "gap-4"
+      }`}>
       {products?.map((product) => (
         <div
           className="relative card bg-white rounded-lg shadow-md overflow-hidden cursor-pointer w-full sm:w-58"
-          key={product.id} 
-        >
+          key={product.id}>
           <div className="relative bg-orange-100">
             {product.image ? (
               <Image
@@ -56,22 +64,36 @@ export default function ProductCard({ products, onDeleteProduct, onDuplicateProd
               <div className="flex flex-wrap justify-center">
                 {buttonsData.map((button, index) => (
                   <button
-                    key={index} 
+                    key={index}
                     onClick={() => button.action(product.id)}
                     className="flex items-center justify-center mb-2 p-2 bg-white rounded-lg hover:bg-gray-200 transition duration-300 mr-2"
-                    style={{ width: '100px' }}
-                  >
-                    <Image src={button.icon} alt={button.text} className="w-5 h-5" />
+                    style={{ width: "100px" }}>
+                    <Image
+                      src={button.icon}
+                      alt={button.text}
+                      className="w-5 h-5"
+                    />
                     <span className="ml-2 text-sm">{button.text}</span>
                   </button>
                 ))}
+                <button
+                  onClick={() => onDownloadProduct(product.image)}
+                  className="flex items-center justify-center mb-2 p-2 bg-white rounded-lg hover:bg-gray-200 transition duration-300 mr-2"
+                  style={{ width: "100px" }}>
+                  <Image
+                    src={designAction.download}
+                    alt="Tải xuống"
+                    className="w-5 h-5"
+                  />
+                  <span className="ml-2 text-sm">Tải Xuống</span>
+                </button>
               </div>
             </div>
           </div>
           <div className="py-4 px-2">
             <h2 className="text-lg font-medium h-20">{product.name}</h2>
           </div>
-        </div>       
+        </div>
       ))}
     </div>
   );
